+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## unreleased (2.2.0-RC3)
+
+### Change
+
+- **Application Requests and Connector Management**
+  - update the status(skipped) in application requests and necessary changes with checkbox in connector management [#1019](https://github.com/eclipse-tractusx/portal-frontend/pull/1019)
+
 ## 2.2.0-RC2
 
 ### Feature